Release/2.5.1

+++ contracts/mock/MockUniswapV2PairTrade.sol
@@ -4,17 +4,16 @@ pragma solidity ^0.8.4;
 import "@uniswap/lib/contracts/libraries/FixedPoint.sol";
 
 contract MockUniswapV2PairTrade {
-
-    uint public price0CumulativeLast;
-    uint public price1CumulativeLast;
+    uint256 public price0CumulativeLast;
+    uint256 public price1CumulativeLast;
 
     uint112 public reserve0;
     uint112 public reserve1;
     uint32 public blockTimestampLast;
 
     constructor(
-        uint _price0CumulativeLast, 
-        uint _price1CumulativeLast, 
+        uint256 _price0CumulativeLast,
+        uint256 _price1CumulativeLast,
         uint32 _blockTimestampLast,
         uint112 r0,
         uint112 r1
@@ -23,22 +22,42 @@ contract MockUniswapV2PairTrade {
         setReserves(r0, r1);
     }
 
-    function getReserves() public view returns(uint112, uint112, uint32) {
+    function getReserves()
+        public
+        view
+        returns (
+            uint112,
+            uint112,
+            uint32
+        )
+    {
         return (reserve0, reserve1, blockTimestampLast);
     }
 
-    function simulateTrade(uint112 newReserve0, uint112 newReserve1, uint32 blockTimestamp) external {
+    function simulateTrade(
+        uint112 newReserve0,
+        uint112 newReserve1,
+        uint32 blockTimestamp
+    ) external {
         uint32 timeElapsed = blockTimestamp - blockTimestampLast;
         if (timeElapsed > 0 && reserve0 != 0 && reserve1 != 0) {
-            price0CumulativeLast += uint(FixedPoint.fraction(reserve1, reserve0)._x) * timeElapsed;
-            price1CumulativeLast += uint(FixedPoint.fraction(reserve0, reserve1)._x) * timeElapsed;
+            price0CumulativeLast +=
+                uint256(FixedPoint.fraction(reserve1, reserve0)._x) *
+                timeElapsed;
+            price1CumulativeLast +=
+                uint256(FixedPoint.fraction(reserve0, reserve1)._x) *
+                timeElapsed;
         }
         reserve0 = newReserve0;
         reserve1 = newReserve1;
         blockTimestampLast = blockTimestamp;
     }
 
-    function set(uint _price0CumulativeLast, uint _price1CumulativeLast, uint32 _blockTimestampLast) public {
+    function set(
+        uint256 _price0CumulativeLast,
+        uint256 _price1CumulativeLast,
+        uint32 _blockTimestampLast
+    ) public {
         price0CumulativeLast = _price0CumulativeLast;
         price1CumulativeLast = _price1CumulativeLast;
         blockTimestampLast = _blockTimestampLast;
